refactor(ff-encode): split video/codec_options.rs into per-codec files
Replace the 766-line monolithic codec_options.rs with a codec_options/
subdirectory containing one file per codec. Each file is under 150 lines.

- codec_options/mod.rs   — VideoCodecOptions enum + re-exports
- codec_options/h264.rs  — H264Options, H264Profile, H264Preset, H264Tune
- codec_options/h265.rs  — H265Options, H265Profile, H265Tier
- codec_options/av1.rs   — Av1Options, Av1Usage
- codec_options/svt_av1.rs — SvtAv1Options
- codec_options/vp9.rs   — Vp9Options
- codec_options/prores.rs  — ProResOptions, ProResProfile
- codec_options/dnxhd.rs   — DnxhdOptions, DnxhdVariant

pub(super) visibility on internal methods changed to pub(in crate::video)
so encoder_inner.rs retains access. Public API is unchanged.

Resolves #709
